chore(nat): Move NF-related tests to nf.rs
In commit 19af92e251e0 ("feat(stateful-nat): move NF to nf.rs") we moved
the code for the NetworkFunction for masquerading NAT to a dedicated
file, but we left the related tests behind; move them along, as it makes
no sense to keep them in the near-empty mod.rs.

Also remove the tracing imports from mod.rs given that there's no more
code in that file.
